+++ src/main/kotlin/com/github/sleepypanda/feesh/settings/categories/Alerts.kt
@@ -35,7 +35,9 @@ object Alerts : CategoryKt("Alerts") {
         this.description = Translated("Shows a title and plays a sound when a rare sea creature is caught by you or your party members. Please enable ${YELLOW}Skyblock Settings -> Personal -> Fishing Settings -> Sea Creature Chat")
     }
 
-    var alertOnSeaCreaturesTypes by select(RareSeaCreatureTypes.CARROT_KING, *RareSeaCreatureTypes.values()) {
+    var alertOnSeaCreaturesTypes by select(
+        *RareSeaCreatureTypes.values().filter { it != RareSeaCreatureTypes.BANSHEE }.toTypedArray(), // Selected by default
+    ) {
         this.name = Translated("Select sea creatures to alert on")
     }