+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Released on: ???
 ## Features
 
 - Added setting to hide other players' fishing hooks.
+- Added Dwarven Lanterns to the deployables tracker and alerts (disabled by default).
 
 ## Bugfixes